requirements.txt
cryptography==44.0.2
python-dotenv
Flask==3.1.2
flask-cors==6.0.0
flask-talisman==1.0.0
future==1.0.0
google-api-python-client==2.133.0
google-cloud-logging
google-cloud-storage==2.17.0
gunicorn==23.0.0
httplib2==0.19.0
jsonschema==2.6.0
oauth2client==4.1.3
pyasn1==0.6.1
PyYAML==6.0.2
pyjwt==2.4.0
python-openid==2.2.5
requests==2.32.4
requests-oauthlib==0.7.0
rsa==4.7
ruamel.yaml==0.18.11
simplejson==3.8.1
six==1.13.0
uritemplate==4.1.1
